requirements.txt
adium-theme-ubuntu==0.3.4
Babel==2.5.1
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cycler==0.9.0
decorator==4.0.6
Django==1.10
django-bootstrap3==9.1.0
django-filter==1.1.0
django-mail-templated==2.6.5
django-password-reset==1.0
django-phonenumber-field==1.3.0
django-registration-redux==1.10
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
duplicity==0.7.6
get==0.0.39
gyp==0.1
idna==2.6
Inshorts==0.2.4
joblib==0.9.4
lockfile==0.12.2
lxml==4.2.0
matplotlib==1.5.1
mechanize==0.2.5
mercurial==3.7.3
nose==1.3.7
numpy==1.14.2
olefile==0.44
pandas==0.22.0
phonenumberslite==8.8.8
Pillow==4.3.0
post==0.0.26
public==0.0.65
pygobject==3.20.0
pyparsing==2.0.3
PySocks==1.6.8
python-dateutil==2.7.0
python-http-client==3.0.0
pytz==2017.3
query-string==0.0.28
request==0.0.26
requests==2.18.4
scikit-learn==0.17
scipy==0.17.0
selenium==3.9.0
sendgrid==5.3.0
simplejson==3.8.1
six==1.11.0
stem==1.6.0
torrequest==0.1.0
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
